[kernel][vm] Correct PmmArena assert
Corrects an assertion in validating the page array allocation of an
arena that was checking if the found range was <= the needed size,
instead of >=. That is to say, it should be considered an error for
the allocated range to not be large enough to hold the desired size.

In practice this is not a bug as boot_reserve_range_search either
returns an error or the exactly passed in requested size.

Change-Id: Id82b633fc6a2d007006766a4ef5667cba0376d16
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1074892
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>