[covargs] Fix bug where multiple embedded build ids causes hang.
This was introduced in fxrev.dev/1074518 because we create a channel
that only has the size = number of profiles, but if there are multiple
embedded build ids in a profile, we'd exceed that size and hang trying
to send more entries to that channel. Instead, we'll use a mutex to just
directly append to the slice of entries.

Bug: 42075455
Change-Id: Ie336ff7763738fbbe7cb6a185e256b73e4d555fc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1075933
Commit-Queue: Ina Huh <ihuh@google.com>
Reviewed-by: Danielle Kay <danikay@google.com>
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>