[virtio-vsock] Use ZX_SOCKET_WRITE_THRESHOLD on Rx
We might be observing a race where the application hasn't read enough
data from the socket by the `1msec` timeout and the connection ends up
in an indefinite stalled state in the current implementation.

Instead of a hardcoded wait, modify a threshold signal for observation
in the zircon socket when needed, so we can update the host with the
available socket space.

Fixed: 300933744
Bug: 345795943
Change-Id: I282b8f917846a5a1c1599a7d5af4ccc68025033f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1069787
Reviewed-by: Christopher Anderson <cja@google.com>
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: David Iserovich <diserovich@google.com>
Commit-Queue: Bruno Dal Bo <brunodalbo@google.com>