[wlan] Adds logging to WLAN state checks in tests
The WLAN policy layer requires that clients acknowledge
receipt of past state updates before serving new updates.
We are seeing some test flakes that suggest that we may not
be synchronizing on this correctly.  This change adds
logging to report the state reports that we are rejecting.

Bug: 350536751
Change-Id: I121535bbbfceb773f5ea59cb9185a53671251633
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1076235
Commit-Queue: Kevin Sakuma <sakuma@google.com>
Reviewed-by: Naomi Mccracken <nmccracken@google.com>