[inspect][benchmarks] Rename fuchsiaperf files so that they're not picked up
Rename the "fuchsiaperf.json" files produced by this test to
"fuchsiaperf.fyi.json" so that they are not picked up by the
performance metrics system.

This is the same as what we did for a similar case of another test:
https://fuchsia-review.googlesource.com/c/fuchsia/+/1066032

These tests produce fuchsiaperf files, but they are not fully hooked
up to the performance metrics system, so the results don't get
uploaded to Chromeperf.

This change should unblock (re-)enabling a check to prevent a
reoccurrence of https://fxbug.dev/343997506.  The check had to be
reverted because we got duplicate results from this test on the
"core.x64-asan-nuc_in_basic_envs" builder, because it got run in both
the NUC and AEMU shards (see https://fxbug.dev/350779634).

Bug: 343997506
Change-Id: I2db9075c4d18927108a32d859e302b39811c70c9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1076867
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Mark Seaborn <mseaborn@google.com>