[botanist] Temporarily increase max attempts for capturing syslog.
We have a max attempts for capturing the syslog within a minute in case
`ffx log` is broken, in which case there's no use continuing trying.
Recently, we've been seeing `ffx log` take a few seconds to start
capturing logs and it fails more than 5 times before that, causing us to
get a "failed to capture syslog" flake. Although this is always
considered a flake at the moment and thus won't fail the build, it does
cause botanist to stop trying to capture the syslog, so we should
temporarily mitigate this by increasing the max attempts until ffx log
can be fixed to not fail so immediately. I also added a new tefmocheck
to fail for the case where we suspect ffx log is broken so we don't miss
it.

Change-Id: I44256c35003cb04e49def6a253e7e6dc872ac034
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1076936
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>