[starnix][memory] Attribute memory for thread groups whose leader is killed
Previously the code relied on `thread_group.leader` being a running
task. But that may not be the case in certain edge cases. The new logic
iterates through all tasks in the thread_group to find the `mm`. In the
vast majority of the cases the iteration should finish in the first loop
since that's the leader.

Fixed: b/349117051
Change-Id: Iebb88aecb21f6561aec4623051cc1a59fb07c9d9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1074475
Commit-Queue: Yifei Teng <yifeit@google.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Reviewed-by: James Robinson <jamesr@google.com>