[starnix][aio] Implement iovec for async I/O
We now implement the READV and WRITEV operations. Also, improve the
error checking for buffer length and offset to pass more gVisor tests.

Skip AIOTest.BasicWrite because this test assumes that the aio context
is a pointer. Our aio context is just a number, which causes the test to
crash. The aio context being a pointer is related to an io ring feature
that we do not implement yet.

Test: BadIOVecs, BadOffset
Bug: b/297456438
Change-Id: Ifb5b50966feccf4b4aa93ba3be8439e2d4b00450
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1076862
Reviewed-by: Benjamin Lerman <qsr@google.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>