[fx][qemu] Removing build logic from `fx qemu`
This removes the custom logic in `fx qemu` that attempts to only
build the artifacts needed to run the emulator.

As we are making changes to the build system, this logic
continues to be fragile and prone to unexpected results. The better
approach is to use the build system and have developers run
`fx build && fx qemu` to make sure all the relevant artifacts are up
to date before starting the emulator.

This also removes the `--no-build` flag since that is now the only
behavior.

Bug: b/346574369
Change-Id: Iab4afe5f2445fb5f756d57d5b06c078ffe7639fc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1077852
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>