[starnix] Introduce MemoryObject
This is currently abstraction over a vmo used by the vfs to provide
memory to map.

This will allow to describe the content of a file with multiple vmos.
This is required to implement the RingBuffer map type for ebpf

Bug: b/323847465
Change-Id: I5f3d3ddc65b4d25270122da5d9cd8ed505e9e183
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1066352
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Adam Perry <adamperry@google.com>