[nand] Optimize copying data
Map VMOs with ZX_VM_MAP_RANGE. This pre-populates all of the page table
entries rather than faulting on them one at a time.

Pre-commit pages to the blobfs transfer buffer. This is necessary for
ZX_VM_MAP_RANGE to work when mapping the VMO.

Don't map the VMO in the NandOperation. Mapping and unmapping the VMO is
slower than 2 vmo_reads or 2 vmo_writes. For blobfs reads, the
oob_buffer isn't set so it's a single vmo_write.

This is an 8% improvement to blobfs page fault times for both compressed
and uncompressed blobs on raw nand.

Bug: b/350020539
Change-Id: I8cd2b1c525b7e9dc9a7e9de9d320f1e395492712
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1075795
Commit-Queue: Chris Drouillard <cdrllrd@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>