[doc-checker] Adding support for reference docs
This adds the feature of being able to check the contents of a reference
documentation directory which is independent of the docs root folder.

This allows generating reference docs and then checking they do not
contain missing or broken links and that they are connected to the
_toc.yaml graph so all reference documentation is reachable from the
table of contents.

Bug: b/344621741
Change-Id: I65131aad9725fcb5a821b310ca99dd032d1d7c40
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1076852
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>