[docs][serdedoc] Adding md_escaped macro in handlebars
This adds a macro to use in the handlebars templates to
escape characters that confuse or break the markdown parsing
of the output.

These characters are things like [] () {}, etc.

This is done in preparation of running doc-checker as part of
the reference doc generation pipeline.

Bug: b/344621741
Bug: b/346408156
Change-Id: I497017c60d19f74ca4c6771c85264c8af07476d8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1077901
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Reviewed-by: Oliver Peralta <ocperalta@google.com>