[flatland] Modify Flatland session instantiation.
Flatland's fidl::Binding object is instantiated on the main thread, not
the one associated with the dispatcher passed as a constructor argument.
This is problematic for two reasons:

1. This interferes with fxr/1072787 (replacing Flatland HLCPP bindings),
   because it causes the natural FIDL bindings to DCHECK.

2. There is an unlikely race condition where the Flatland instance
   processes a message before the instance has been fully constructed.
   (AFAIK this condition has never been observed).

This CL addresses both issues by moving the instantiation of the binding
onto the Flatland instance's dispatcher/thread.

Bug: 350823293
Change-Id: I439d07123f2f885da377258aa0cdc1ef9ef9cdf8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1076398
Commit-Queue: Josh Gargus <jjosh@google.com>
Reviewed-by: Emircan Uysaler <emircan@google.com>