[bt][avrcp] Fix avrcp relay incomplete setup issue.
Sometimes, peers do not respond to our initial AVRCP requests to get the
media attributes/play status and this causes us to hang and not send a
WatchInfoChange response to mediasession.
Implement a work around so that place a time-out on the requests,
so that it doesn't block us from setting up the relay.

Test: fx test bt-a2dp-tests
Bug: b/343196880
Change-Id: I72754b6818923aa8867f3989f703714e04454865
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1073538
Reviewed-by: Ani Ramakrishnan <aniramakri@google.com>
Commit-Queue: Dayeong Lee <dayeonglee@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>