[driver_runtime] Arena & dispatcher rust bindings
This adds implementations of safe rust bindings for the arena and
dispatcher parts of the driver runtime api surface. This allows for the
following in the public api of the crate:

- Creating and destroying arenas
- allocating arena lifetime-constrained memory from an arena that drop
  appropriately when they go out of scope
- posting tasks to a running dispatcher

It also allows for creating a top level driver runtime dispatcher to run
tests under in the crate-internal api, and uses that to exercise the
driver runtime async executor driving rust futures to completion.

Multiply: fdf_rust_tests
Fixed: b/338249591
Change-Id: Ib69da7b35187ac6f4ec6d550d27b5f848655cd0b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1070133
Reviewed-by: David Koloski <dkoloski@google.com>
Reviewed-by: Charles Celerier <chcl@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>
Reviewed-by: Novin Changizi <novinc@google.com>