[ffx]Removing unreachable help ffx_plugin
This removes the ffx_plugin for Help in the ffx frontend.

It is unreachable since argh handles the "help" subcommand.
The jinja template used to chain together all the plugins and subtools
still expects the top level of ffx to be a "plugin suite", even if
the code is unreachable, so some boilerplate is needed until the
jinja template is edited or removed.

This is the last instance of ffx_core::ffx_plugin.

Bug: b/290067014
Change-Id: Ie94653189300b4e83771187140ae900aed9a7133
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1078795
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>