[fuchsia.hardware.network] Deprecate DeviceClass
... in favor of a new flexible enum `PortClass`. This will ensure the
FIDL types can evolve as Fuchsia adds support for new classes of ports.

All internal APIs have been migrated to use the new `PortClass`. The
fuchsia.net.interfaces API, however, is published to the SDK, so its
`Properties` struct contains both `device_class` (old) and `port_class`
(new). Once support for `DeviceClass` is officially retired, this can
be cleaned up. It is therefore the responsibility of the netstack to
populate both fields; an integration test covering this has been added.

The majority of the "real changes" in this CL is introducing error
handling in the components that interact with the `PortClass`, and now
need to handle the unknown variant gracefully. The largest changes from
this occurred in Netstack/Netcfg.

While here, introduce the `Lowpan` port variant, which was the
motivating case for making the enum flexible. The lowpan component
now specifies this class when creating the interfaces. Notably, a TODO
in MDNS could be cleaned up, as they were previously searching for the
lowpan interface by name.

Bug: 42157740
Fixed: 349436251
Change-Id: I95ac25cfdc9d27b2435c090ce7cc1b8493a07cb3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1073292
Reviewed-by: Sam Balana <sbalana@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
API-Review: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Will Drewry <drewry@google.com>
Reviewed-by: Colin Nelson <colnnelson@google.com>
Reviewed-by: Prashanth Swaminathan <prashanthsw@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>