[kernel][elfldltl] Complicate static_assert to avoid tautological test
GCC 14's "tautological test" warning triggers on trivial
expressions like "X != X" even when X has a template-dependent
value and the whole thing is never evaluated because it's in an
uninstantiated template or an `if constexpr` branch not taken.
Making the expression for constant false more nontrivially
template-dependent avoids this.  The easy mechanism is to use the
non-type template parameter as a parameter in another template
class and test that with `std::is_void`; it will always be false
since no template class instantiation can be the void type, but
the compiler won't really clock that because there could be an
explicit specialization for some `std::is_void<T>` where T is
template-dependent, and it doesn't know there won't ever be one.

Bug: 351904809
Change-Id: Ic74f7a1f940684c64a0a85f76926947101b2d09f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1078587
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: John Grossman <johngro@google.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>