[devicetree][gpio] Make gpio-init bind to a specific controller
Make composite node specs with a gpio-init parent bind to the specific
GPIO controller that is responsible for executing the init steps that
the composite node depends on. This ensures that the initialization
order is correct on boards with multiple GPIO controllers.

Test: fx test gpioimpl-visitor-test
Change-Id: Id427b3060a352d604926ad2e9fd951f6c886b2a3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1060144
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Novin Changizi <novinc@google.com>
Fuchsia-Auto-Submit: Braden Kell <bradenkell@google.com>
Reviewed-by: Christopher Anderson <cja@google.com>