[starnix][pty] Send signals to the correct process group
We are only supposed to generate signals when transforming the terminal
input queue and those signals are always delivered to the foreground
process group of the controlling session for the terminal.

Previously, we computed signals both for the input and output queues,
and we maintained two session associations, one for each end of the
terminal. None of that complexity appears to be correct. This CL
simplifies these data structures and passes more tests.

Additionally, the terminal was maintaining its own concept of the
foreground process group for a session, which appears to be incorrect.
The notion of a foreground process group is scoped to the session
itself, not to a given terminal.

Test: JobControlTest.ControllingProcessPersistsAfterChildExists
Fixed: b/42159560
Change-Id: I201a0cbd7b1ee935864d324d46ac5009901984d0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1078792
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Benjamin Lerman <qsr@google.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>