[superproject] Roll scudo [scudo] Add a maximum value into the timer. (#96989)
Modify the timer code to keep track of the maximum time that any call
takes.

Add a method to get a ScopedString representing the timer data. Allows
the testing of the timer on all platforms.

Add new unit tests for the maximum, and a lot of new tests.

GitOrigin-RevId: 543b5b7ad21f5a8a783064b67a1a458db1ef71ae
Original-Revision: 9bf88109ef5c887cf52948ed34d45e08cc1695f8
Roller-URL: https://ci.chromium.org/b/8742720036711361345
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I5585900e8ede0a7c9550c57a457b7f5f03a8b35f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1079478