[sdk][build] Only allow sdk_fuchsia_package() in default_toolchain
Previously, the template supported non-default toolchains by
including extra logic and creating an empty group. This added
unnecessary complexity and could be applied inconsistently depending on
the value of `package_in_target_api_level`.

Instead, only allow SDK packages to be created for the
`default_toolchain` and require users of the template to ensure it is
only used with this toolchain. This is consistent with
https://fuchsia.dev/fuchsia-src/development/build/build_system/best_practices_toolchains.
(Note: `default_toolchain` is equivalent to `target_toolchain`, which
was previously used.)

Update GN files that use this template and  get processed with other
toolchains (`fidling` and the `-shared` variant) to avoid using the
template with those toolchains.

Also print a message in the only remaining alternate handling, a noop
package when the package is not supported in the target API level. This
can inform developers working in this area that a package is being
excluded while not adding build spam because sub-build output is not
sent to stdout by default.

Bug: 347094435
Change-Id: Ie958830670390d918c5351a72e1dc188468c7171
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1058722
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: David Dorwin <ddorwin@google.com>