[debugger] Improve help text when failing to infer filenames
The previous text was not particularly helpful:

```
[zxdb] ...
There is no current file name to use, you'll have to specify a file.
```

Now we give better instructions for what to do:

```
[zxdb] ...
There is no file name to use in the current context. Either specify a
file and line number (e.g. main.cc:23), or ensure the process is stopped (try
`pause`) and the file that you want to stop in is the active stack frame with
`frame` to use line numbers with a contextual file name.

See `help break` for details.
```

Change-Id: Ib635db7c311c219134af170afa169a42fffaf751
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1079632
Fuchsia-Auto-Submit: Jacob Rutherford <jruthe@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>