[starnix][gvisor] Skip AIOTest.BadWrite
This test is flaky because we generate the error that the test is
looking for asynchronously, which means the test passes when we're fast
enough to generate the error before the test checks for it.

The real fix is to rework our async I/O code to do more work
synchronously, but, for now, we just skip the test to avoid the flake.

Fixed: b/352233059
Change-Id: I0465ddf5717c14bc5370baa5fe6ea0c4139c379f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1080132
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>