Reland "[dtr] realm factory use test.Capability"
This is a reland of commit 2bcb4080d7570578f27edef5513f1f6a97f6483e

There used to be an entry for driver_urls in the driver test realm's
RealmArgs table at entry 14. This was removed as it was not part of the
stable sdk. Unfortunately we have CTF tests that are frozen that ended
up using that field. This causes the decoding on the realm factory side
to break. Disabling these tests until those branches are updated to
remove the driver_url field so that the tests can be reenabled.

Original change's description:
> [dtr] realm factory use test.Capability
>
> In the driver test realm's start args, we can re-use the
> fuchsia.component.test.Capability instead of creating our own type. This
> will allow us to offer and expose other capabilities instead of being
> tied to the offer=protocol and expose=service. This also changes how
> these are routed in the driver test realm, to directly forward these
> instead of manually serving up exposes. Keeps the old way for a soft
> migration, but marks them as deprecated. These new fields are added into
> the parner sdk.
>
> Tests: basic-driver-test
> Bug: 42073139
>
> Change-Id: Ifb4de4cd5659fe442a2b99d5f5d13b09f0b54d13
> Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1061660
> API-Review: Jocelyn Dang <jocelyndang@google.com>
> Reviewed-by: Jocelyn Dang <jocelyndang@google.com>
> Fuchsia-Auto-Submit: Novin Changizi <novinc@google.com>
> Reviewed-by: Sarah Chan <spqchan@google.com>
> Commit-Queue: Novin Changizi <novinc@google.com>

Bug: 42073139
Change-Id: Id4dc10c2120e4fde650dca28ff87ffa717c1f060
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1079512
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Sarah Chan <spqchan@google.com>
API-Review: Jocelyn Dang <jocelyndang@google.com>
Commit-Queue: Novin Changizi <novinc@google.com>