[async_utils] Support long lines in command output
It turns out the internal buffer for StreamReader is only 64KiB, and
when a single line is longer than the limit an error is raised. Since we
are running in a task using asyncio.wait(), the exception is not
propagated and the program hangs.

This change manually processes outputs from the StreamReader without
using the internal buffer, and it additionally ensures that exceptions
in any read or write tasks are propagated up.

Change-Id: I87e2bd349e5460847d4080042c82e1595419cd1a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1080074
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Fuchsia-Auto-Submit: Christopher Johnson <crjohns@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>