[debugger] Fix name resolution after dot operator
In specific situations where there is a type that conflicts with a
member name, the expression parser will always take the type before
looking for variable identifiers with that name.

In the context of evaluating the right hand side of a "." or "->"
operator in an expression, types should be ignored. If a type does
happen to conflict with the token that would otherwise identify a
particular member on an object, we should treat the name of the type
like an identifier and then let the member accessor logic determine if
there is actually an identifier there by that name.

Fixed: 352125964
Change-Id: I79538c6c82a8dbe23de58fbf95cc77563a419bf9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1079633
Commit-Queue: Jacob Rutherford <jruthe@google.com>
Reviewed-by: Adam Barth <abarth@google.com>