[tee] Fork tee-client-api
This copies the code as-is from //src/security/lib/tee/tee-client-api.
This is intended to allow for experimentation in cases where the client
and TA are both running in the same Fuchsia environment, without risking
breakage of the existing tee-client-api.

At this commit, this fork modifies the original contents by:
-Omitting OWNERS and METADATA.textproto
-Updating the README
-Updating headers via fx format-code
-Renaming the build targets and source file to tee-client-api-experimental

Bug: b/352145137
Change-Id: I264e57f5396b419396fb88e8895050f024cd61fd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1080114
Fuchsia-Auto-Submit: Eric Stone <ecstone@google.com>
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Eric Stone <ecstone@google.com>
Reviewed-by: Mike Comfoltey <comfoltey@google.com>