[power][broker] Create internal lease element
Adds an intermediate, synthetic lease element that has an assertive
dependency to the element being leased. This changes the behavior of
leasing elements in the power broker to return 'Satisfied' only when the
element being leased has reached the desired level. Prior to this
change, the broker would return 'Satisfied' when the dependencies of the
leased element have reached their required levels.

Bug: 345829869
Test: fx test src/power/broker
Test: fx test fake-system-activity-governor-test
Change-Id: I4ea1dbf58dda2bd778117bcaf69d1ebc5f62776d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1066613
Reviewed-by: Onath Dillinger <claridge@google.com>
Reviewed-by: Michael Brunson <mbrunson@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Caroline Liu <carolineliu@google.com>
Commit-Queue: Prashanth Swaminathan <prashanthsw@google.com>