[netstack3] Expose new many tx allocation fn
A previous CL introduced a function to allocate more than one TX buffer
at once, but failed to expose it on the crate's public API.

In order to allow a borrowed method from Session, it must carry a strong
reference to the inner state, which defeats the Task not running
mechanism. It turns out that mechanism is very racy and not offering
great protection, so we remove it altogether.

Bug: 42057204
Change-Id: I99248225c887d4172d8200ae7e198986aa1504ca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1080634
Reviewed-by: Zeling Feng <zeling@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>