[netstack3] Simplify DeviceSendFrame error
In preparation to better handling of dequeueing frames, simplify the
errors that can be observed when sending frames to the device.

First, reframe the generic "NotReady" error (which is never yielded by
bindings) to "NoBuffers" which is closer to errors that we might
actually observe in practice -- though we're not yielding that just yet
because the tx queue handler in bindings is not going to behave nicely
if it observes that error.

If any errors are observed during dequeueing, we just drop the packet
and increment a counter. Given this packet has already been delivered to
packet sockets we should not be requeueing it again.

Bug: 42057204
Change-Id: I80cd058faf8875888c65ee6db159c11d912f81e7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1080212
Reviewed-by: Evan Wildenhain <wildenhain@google.com>
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>