[driver_runtime] Remove rust `Task::cancel` for now
We don't actually need this yet and there are problems with the way it
works making it hard to properly test. Removing this without completely
removing the mechanisms that help make it possible so that it can be
re-added in the future when a more resilliant design can be made.

Change-Id: I0d3a13dfd362212b00cba0f5338c3b0cbc95e760
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1080433
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>
Reviewed-by: Charles Celerier <chcl@google.com>
Reviewed-by: Novin Changizi <novinc@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>