[fxfs] Use bloom filters for queries
Adds a bloom filter to the persistent layer file format which allows us
to quickly determine whether a layer file needs to be consulted for a
given search.

To make use of the bloom filters, the LSM tree needs to know what the
goal of the caller is when performing an `LsmTree::seek`.  Added a
`Query` type which expresses the goal and allows us to use the bloom
filter appropriately; see lsm_tree/merge.rs for details.

Bug: 333401205
Change-Id: I1d2e7ecb7abb4c5ee9a8875e1f626df3e435c8f7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1017312
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Chris Suter <csuter@google.com>