[idk][build] Remove sdk_molecule's host_tools_deps
https://fxrev.dev/943927 introduced the `host_tools_deps` argument to
the `sdk_molecule()` template as a means to remove the global
`sdk_no_host_tools` arg. However, not all targets assigned to it are
host tools, it is only used once per IDK, and it hides a decision
about what is included in the IDK build.

Instead, let the IDKs specify when to include such dependencies. In the
future, we will be able to use other GN variables and remove the global
GN arg `sdk_inside_sub_build`.

Bug: 347094435
Change-Id: I917eeec2b9773a45c127a27f951fa0ffd11b9624
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1074468
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: David Dorwin <ddorwin@google.com>