[versioning] Add support for NEXT to version_history
Add support for API level "NEXT" to version_history scripts and code so
that they are ready for the switch to "NEXT" as the only in-development
API level.

get_platform_version.py now requires that `api_levels` have a status of
"unsupported", "sunset", "supported", or "in-development", and that
`special_api_levels` have a status of "in-development". This was already
the case in practice but was not enforced.

Bug: 326277078
Change-Id: I1e8356bfebbfda6dca61bf468dfcb9283f7c357d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1073565
Commit-Queue: David Dorwin <ddorwin@google.com>
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
Reviewed-by: David Fang <fangism@google.com>
Fuchsia-Auto-Submit: David Dorwin <ddorwin@google.com>