[vfs][rust] Return EntryInfo for Node and DirectoryEntry
Introduce a new trait `GetEntryInfo` that all implementations of
`DirectoryEntry` (used in pseudo filesystems) and `Node` must
implement. This trait only has one method that returns `EntryInfo`.

Prior to this change, there is a method to get `EntryInfo` but it was
a part of only `DirectoryEntry`. Being able to get the entry's
information, regardless of if it is used in a pseudo filesystem, is
still useful. For example, in Open3, we want to identify the directory
entry type to validate protocols when there are more than one specified.

Bug: 348698584
Test: fx test //src/storage/conformance
Change-Id: Idefee013e4e96e8803c92d2166a15f27fe2d23b9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1079792
Reviewed-by: James Sullivan <jfsulliv@google.com>
Reviewed-by: Brandon Castellano <bcastell@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Owners-Override: Adam Barth <abarth@google.com>