[intel-i915] Pass an Inspector on Controller creation.
This change adds an Inspector argument to the Controller constructor and
its factory function. This makes it possible for drivers to pass an
Inspect to a hardware component class, which is required for DFv2
drivers to set up the component inspector.

Test: intel-i915-test
Test: On core.x64 on NUC11, "ffx target snapshot" output contains
      intel-i915 root node.
Change-Id: I67882c087102acc996c337920529d625a29ee00e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1076537
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Victor Costan <costan@google.com>