[netstack3] Move Segment into base
...along with all transitive dependencies.

We want to be able to reuse this struct as part of conntrack's TCP state
tracking. It couldn't be used directly, since the netstack3_tcp crate
has a dependency on the filtering crate, so we'd end up with a
dependency cycle.

This also required moving a few impl blocks to avoid running afoul of
the orphan rule.

I tried to retain history as much as possible, but a few files were only
partially copied, so their history is sadly lost.

Bug: 328064736
Change-Id: I135c497663c3ba670648d73423caa60928d44e3a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1080553
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Peter DeLong <pdelong@google.com>