[netstack3] Don't spawn a TX task for loopback
We're spawning this task but the loopback interface is not actually
configured to use any queues. This is just wasting memory and increasing
complexity. We can reintroduce the TX task when we add queueing
discipline for loopback.

To make the lack of support clear, don't create a tx notifier for
loopback, which will make the stack panic loudly if we ever configure
loopback to have a tx queue.

Right now this is just getting in the way of a better tx queue
implementation for netdevice-backed interfaces.

Bug: 42057204

Change-Id: Idcef9871172f8e1460e69e9fef085186666ad22f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1080726
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>